Add amount range to `lease_rate`
We add a minimum and maximum amount to the lease rate. This allows
sellers to provide different rates depending on the amount that is
purchased. Since the number of utxos used will likely vary based
on the amount, this also lets them use more accurate funding weights
to ensure they don't end up paying too much of the on-chain fees.

With that data added, a lease rate uses 24 bytes. We limit the number
of rates in a `node_announcement` to 10, which doesn't make it too
harmful in terms of bandwidth usage for the gossip network.